feat: 전체 카테고리 조회
