--- src/main/java/gift/model/category/CategoryService.java
@@ -1,22 +1,0 @@
-package gift.model.category;
-
-import gift.repository.CategoryRepository;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@Transactional(readOnly = true)
-public class CategoryService {
-
-    private final CategoryRepository categoryRepository;
-
-    public CategoryService(CategoryRepository categoryRepository) {
-        this.categoryRepository = categoryRepository;
-    }
-
-    @Transactional
-    public CategoryResponse register(CategoryRequest categoryRequest) {
-        Category category = categoryRepository.save(categoryRequest.toEntity());
-        return CategoryResponse.from(category);
-    }
-}
