+++ src/main/java/gift/repository/CategoryRepository.java
@@ -1,8 +1,12 @@
 package gift.repository;
 
 import gift.model.category.Category;
+import gift.model.product.Product;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CategoryRepository extends JpaRepository<Category, Long> {
 
+    Page<Category> findAll(Pageable pageable);
 }
